+++ test/utility_test.cpp
@@ -1,11 +1,13 @@
 /**
  * Tests for the utility.h header. Includes helper methods for dealing with
  * sparse matrices.
+ *
+ * Copyright 2022 Peter Macgregor
  */
 #include <gtest/gtest.h>
 #include <utility.h>
 
-TEST(StagTest, IsSymmetric) {
+TEST(UtilityTest, IsSymmetric) {
   // Construct a symmetric matrix.
   std::vector<int> rowStarts = {0, 2, 4, 7, 8};
   std::vector<int> colIndices = {1, 2, 0, 2, 0, 1, 3, 2};